move filters.py to python module
