+++ backend/corpora/dbnl/dbnl.py
@@ -8,7 +8,7 @@ from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
 from addcorpus.extract import Metadata, XML, Pass, Order, Backup, Combined
 import corpora.dbnl.utils as utils
 from addcorpus.es_mappings import *
-from addcorpus.filters import RangeFilter, MultipleChoiceFilter, BooleanFilter
+from addcorpus.python_corpora.filters import RangeFilter, MultipleChoiceFilter, BooleanFilter
 from corpora.dbnl.dbnl_metadata import DBNLMetadata
 
 class DBNL(XMLCorpusDefinition):
