+++ backend/corpora/ecco/ecco.py
@@ -11,7 +11,7 @@ import re
 from django.conf import settings
 
 from addcorpus.extract import Combined, Metadata, XML
-from addcorpus import filters
+from addcorpus.python_corpora import filters
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
 from addcorpus.es_settings import es_settings
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
