+++ backend/corpora/goodreads/goodreads.py
@@ -9,7 +9,7 @@ from django.conf import settings
 import openpyxl
 
 from addcorpus.extract import CSV, Metadata
-from addcorpus.filters import MultipleChoiceFilter, RangeFilter
+from addcorpus.python_corpora.filters import MultipleChoiceFilter, RangeFilter
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition, FieldDefinition
 
 from addcorpus.es_mappings import main_content_mapping
