+++ backend/corpora/guardianobserver/guardianobserver.py
@@ -16,7 +16,7 @@ from django.conf import settings
 
 from es.es_update import update_document
 from addcorpus import extract
-from addcorpus import filters
+from addcorpus.python_corpora import filters
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition, until, after, string_contains, consolidate_start_end_years
 from media.image_processing import sizeof_fmt
 from media.media_url import media_url
