+++ backend/corpora/parliament/finland-old.py
@@ -3,7 +3,7 @@ from glob import glob
 
 from addcorpus.python_corpora.corpus import CSVCorpusDefinition
 from addcorpus.extract import CSV, Combined, Constant
-from addcorpus.filters import MultipleChoiceFilter
+from addcorpus.python_corpora.filters import MultipleChoiceFilter
 from corpora.parliament.parliament import Parliament
 import corpora.parliament.utils.field_defaults as field_defaults
 from corpora.utils.constants import document_context
