+++ backend/corpora/periodicals/periodicals.py
@@ -13,7 +13,7 @@ import openpyxl
 from django.conf import settings
 
 from addcorpus import extract
-from addcorpus import filters
+from addcorpus.python_corpora import filters
 from addcorpus.python_corpora.corpus import XMLCorpusDefinition, FieldDefinition
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
 from addcorpus.es_settings import es_settings
