+++ backend/corpora/rechtspraak/rechtspraak.py
@@ -8,10 +8,11 @@ from zipfile import ZipFile, BadZipFile
 
 from django.conf import settings
 
-from addcorpus import extract, filters
+from addcorpus import extract
 from addcorpus.python_corpora.corpus import FieldDefinition, XMLCorpusDefinition
 from addcorpus.es_mappings import keyword_mapping, main_content_mapping
 from addcorpus.es_settings import es_settings
+from addcorpus.python_corpora import filters
 
 logger = logging.getLogger('indexing')
